+++ CHANGELOG.md
@@ -10,11 +10,14 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+## [1.1.0] - 2024-04-19
+
 - Improve support for SUSE/Arch based containers/self-hosted runners.
 
 ## [1.0.0] - 2024-03-08
 
 Initial release
 
-[Unreleased]: https://github.com/taiki-e/checkout-action/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/taiki-e/checkout-action/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/taiki-e/checkout-action/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/taiki-e/checkout-action/releases/tag/v1.0.0